requirements.txt
python-dotenv
pandas
numpy
chardet
nbformat
nbconvert
requests
lxml
fsspec
rest-pandas
pymongo
djongo
mongoengine